Highlight lines with modified expand region
Select a number of rows for copying or highlighting with a simple wrapper around
Magmar's expand region. This changes the behavior when you add a numeric prefix,
for instead of guessing at how many levels of context to surround, this
surrounds that many lines (just seems a little more intuitive to me).